refactored and added a subpage for each lecture
